fixing reminders
